Made sprite playAnimation better and moved chunk/map
